Update SQLDebugTool.html
Rewrote copytoclipboard function as it was only working over https. 
navigator.clipboard API is not available in the context in which the script is running. 

This API is only available in secure browsing contexts (HTTPS), and also requires the user to grant permission. If the page is being served over HTTP or if permissions are not granted, navigator.clipboard will be undefined.

If navigator.clipboard is not available, the function falls back to using document.execCommand